Fix bug in finding value implementation
Previously, it would search for a value in the complete newline, irrespective of the previous
key that was found in the previous level. If there is a value like {"a":{},"b":{"c":""}}, it
would find a value for $.a.c because there is a "c" key on the correct level. This fixes it
by limiting the search to the end of the level.